+++ craft_cli/dispatcher.py
@@ -281,7 +281,7 @@ class Dispatcher:
         return ArgumentParsingError(self._help_builder.get_usage_message(text))
 
     def _get_requested_help(  # noqa: PLR0912 (too many branches)
-        self, parameters: list[str]
+        self, parameters: list[str], app_config: Any
     ) -> str:
         """Produce the requested help depending on the rest of the command line params."""
         if len(parameters) == 0:
@@ -332,7 +332,7 @@ class Dispatcher:
             raise self._build_usage_exc(msg) from None
 
         # instantiate the command and fill its arguments
-        command = cmd_class(None)
+        command = cmd_class(app_config)
         parser = _CustomArgumentParser(self._help_builder, prog=command.name, add_help=False)
         command.fill_parser(parser)
 
@@ -413,7 +413,7 @@ class Dispatcher:
                 filtered_sysargs.append(sysarg)
         return global_args, filtered_sysargs
 
-    def pre_parse_args(self, sysargs: list[str]) -> dict[str, Any]:
+    def pre_parse_args(self, sysargs: list[str], app_config: Any = None) -> dict[str, Any]:
         """Pre-parse sys args.
 
         Several steps:
@@ -423,6 +423,8 @@ class Dispatcher:
         - validate global options and apply them
 
         - validate that command is correct (NOT loading and parsing its arguments)
+
+        If provided, ``app_config`` is passed to the command to be validated.
         """
         global_args, filtered_sysargs = self._parse_options(self.global_arguments, sysargs)
 
@@ -448,7 +450,7 @@ class Dispatcher:
 
         # handle requested help through -h/--help options
         if global_args["help"]:
-            help_text = self._get_requested_help(filtered_sysargs)
+            help_text = self._get_requested_help(filtered_sysargs, app_config)
             raise ProvideHelpException(help_text)
 
         if not filtered_sysargs or filtered_sysargs[0].startswith("-"):
@@ -466,7 +468,7 @@ class Dispatcher:
 
         # handle requested help through implicit "help" command
         if command == "help":
-            help_text = self._get_requested_help(cmd_args)
+            help_text = self._get_requested_help(cmd_args, app_config)
             raise ProvideHelpException(help_text)
 
         self._command_args = cmd_args